Refactor; delete useless test code.
