+++ integration_tests/src/test/spark320/scala/com/nvidia/spark/rapids/shims/OrcStatisticShim.scala
@@ -30,7 +30,7 @@
 {"spark": "340"}
 {"spark": "341"}
 spark-rapids-shim-json-lines ***/
-package com.nvidia.spark.rapids.shims.spark311
+package com.nvidia.spark.rapids.shims
 
 import java.util.Objects
 
@@ -38,10 +38,10 @@ import org.apache.orc._
 
 object OrcStatisticShim {
   def supports(left: ColumnStatistics, right: ColumnStatistics): Boolean = (left, right) match {
-    case (dateStat: DateColumnStatistics, otherDateStat: DateColumnStatistics) => true
-    case (strStat: StringColumnStatistics, otherStrStat: StringColumnStatistics) => true
-    case (cStat: CollectionColumnStatistics, otherCStat: CollectionColumnStatistics) => true
-    case (l@_, r@_) => false
+    case (_: DateColumnStatistics, _: DateColumnStatistics) => true
+    case (_: StringColumnStatistics, _: StringColumnStatistics) => true
+    case (_: CollectionColumnStatistics, _: CollectionColumnStatistics) => true
+    case _ => false
   }
 
   def equals(left: ColumnStatistics, right: ColumnStatistics): Boolean = (left, right) match {
